Removing Inner structs in cpp client
