+++ pulsar-client-cpp/include/pulsar/CompressionType.h
@@ -1,0 +1,27 @@
+/**
+ * Copyright 2016 Yahoo Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef PULSAR_COMPRESSIONTYPE_H_
+#define PULSAR_COMPRESSIONTYPE_H_
+namespace pulsar {
+enum CompressionType {
+    CompressionNone = 0,
+    CompressionLZ4 = 1,
+    CompressionZLib = 2
+};
+}
+
+#endif /* PULSAR_COMPRESSIONTYPE_H_ */
