+++ pulsar-client-cpp/lib/Client.cc
@@ -30,142 +30,6 @@ DECLARE_LOG_OBJECT()
 
 namespace pulsar {
 
-struct ClientConfiguration::Impl {
-    AuthenticationPtr authenticationPtr;
-    AuthDataPtr authDataPtr;
-    int ioThreads;
-    int operationTimeoutSeconds;
-    int messageListenerThreads;
-    int concurrentLookupRequest;
-    std::string logConfFilePath;
-    bool useTls;
-    std::string tlsTrustCertsFilePath;
-    bool tlsAllowInsecureConnection;
-    Impl() : authenticationPtr(AuthFactory::Disabled()),
-             authDataPtr(Auth::Disabled()),
-             ioThreads(1),
-             operationTimeoutSeconds(30),
-             messageListenerThreads(1),
-             concurrentLookupRequest(5000),
-             logConfFilePath(),
-             useTls(false),
-             tlsAllowInsecureConnection(true) {}
-};
-
-ClientConfiguration::ClientConfiguration()
-        : impl_(boost::make_shared<Impl>()) {
-}
-
-ClientConfiguration::~ClientConfiguration() {
-}
-
-ClientConfiguration::ClientConfiguration(const ClientConfiguration& x)
-    : impl_(x.impl_) {
-}
-
-ClientConfiguration& ClientConfiguration::operator=(const ClientConfiguration& x) {
-    impl_ = x.impl_;
-    return *this;
-}
-
-ClientConfiguration& ClientConfiguration::setAuth(const AuthenticationPtr& authentication) {
-    impl_->authenticationPtr = authentication;
-    return *this;
-}
-
-const Authentication& ClientConfiguration::getAuth() const {
-    return *impl_->authenticationPtr;
-}
-
-ClientConfiguration& ClientConfiguration::setAuthentication(const AuthDataPtr& authentication) {
-    impl_->authDataPtr = authentication;
-    impl_->authenticationPtr = AuthData::getAuthenticationPtr(authentication);
-    return *this;
-}
-
-const AuthData& ClientConfiguration::getAuthentication() const {
-    return *(impl_->authDataPtr);
-}
-
-
-const AuthenticationPtr& ClientConfiguration::getAuthenticationPtr() const {
-    return impl_->authenticationPtr;
-}
-
-ClientConfiguration& ClientConfiguration::setOperationTimeoutSeconds(int timeout) {
-    impl_->operationTimeoutSeconds = timeout;
-    return *this;
-}
-
-int ClientConfiguration::getOperationTimeoutSeconds() const {
-    return impl_->operationTimeoutSeconds;
-}
-
-ClientConfiguration& ClientConfiguration::setIOThreads(int threads) {
-    impl_->ioThreads = threads;
-    return *this;
-}
-
-int ClientConfiguration::getIOThreads() const {
-    return impl_->ioThreads;
-}
-
-ClientConfiguration& ClientConfiguration::setMessageListenerThreads(int threads) {
-    impl_->messageListenerThreads = threads;
-    return *this;
-}
-
-int ClientConfiguration::getMessageListenerThreads() const {
-    return impl_->messageListenerThreads;
-}
-
-ClientConfiguration& ClientConfiguration::setUseTls(bool useTls) {
-    impl_->useTls = useTls;
-    return *this;
-}
-
-bool ClientConfiguration::isUseTls() const {
-    return impl_->useTls;
-}
-
-ClientConfiguration& ClientConfiguration::setTlsTrustCertsFilePath(const std::string &filePath) {
-    impl_->tlsTrustCertsFilePath = filePath;
-    return *this;
-}
-
-std::string ClientConfiguration::getTlsTrustCertsFilePath() const {
-    return impl_->tlsTrustCertsFilePath;
-}
-
-ClientConfiguration& ClientConfiguration::setTlsAllowInsecureConnection(bool allowInsecure) {
-    impl_->tlsAllowInsecureConnection = allowInsecure;
-    return *this;
-}
-
-bool ClientConfiguration::isTlsAllowInsecureConnection() const {
-    return impl_->tlsAllowInsecureConnection;
-}
-
-ClientConfiguration& ClientConfiguration::setConcurrentLookupRequest(int concurrentLookupRequest) {
-    impl_->concurrentLookupRequest = concurrentLookupRequest;
-    return *this;
-}
-
-int ClientConfiguration::getConcurrentLookupRequest() const {
-    return impl_->concurrentLookupRequest;
-}
-
-ClientConfiguration& ClientConfiguration::setLogConfFilePath(const std::string& logConfFilePath) {
-    impl_->logConfFilePath = logConfFilePath;
-    return *this;
-}
-
-const std::string& ClientConfiguration::getLogConfFilePath() const {
-    return impl_->logConfFilePath;
-}
-
-/////////////////////////////////////////////////////////////////
-
 Client::Client(const std::string& serviceUrl)
         : impl_(boost::make_shared<ClientImpl>(serviceUrl, ClientConfiguration(), true)) {
 }
